move rom loader code to its own file
